fix(action): sanity yamllint output
When running in GitHub workflow, the yamllint outputs are being grouped
via GitHub using `::group::` and `::endgroup::`.
This grouping is also added to the output variable, but as it's not a
valid command for setting messages (see for example setting a warning
message in GitHub docs), this grouping needs to be removed.

https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#grouping-log-lines
https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-a-warning-message